% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{aar_ktr_gr}
\alias{aar_ktr_gr}
\title{lage aarskontrollfigur}
\usage{
aar_ktr_gr(df, k)
}
\arguments{
\item{df}{data frame}

\item{k}{which year control}
}
\value{
df data frame grouped by year and hospital
}
\description{
lage aarskontrollfigur
}
