% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{aar_ktr_tb}
\alias{aar_ktr_tb}
\title{lage aarskontrolltabell}
\usage{
aar_ktr_tb(df, k = 2)
}
\arguments{
\item{df}{data frame}

\item{k}{which year control}
}
\value{
df data frame grouped by year and hospital
}
\description{
lage aarskontrolltabell
}
