% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{aarKtrl}
\alias{aarKtrl}
\title{lage aarskontrolltabell}
\usage{
aarKtrl(df, k, agg)
}
\arguments{
\item{df}{data frame}

\item{k}{which year control}

\item{agg}{lgl resultater aggregert}
}
\value{
df data frame grouped by year and hospital
}
\description{
lage aarskontrolltabell
}
