% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aarsktrl.R
\name{aars_ktrl}
\alias{aars_ktrl}
\title{Pick particular hospitals and years from a data frame}
\usage{
aars_ktrl(df)
}
\arguments{
\item{df}{Data frame holding full SoReg data}
}
\value{
k1nt A data frame for choice of hospitals and years

k2nt A data frame for choice of hospitals and years
}
\description{
Pick particular hospitals and years from a data frame
}
