% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{detail}
\alias{detail}
\title{lage vekttapdetaljer}
\usage{
detail(dm)
}
\arguments{
\item{df}{data frame}

\item{k}{which year control}
}
\value{
df data frame grouped by year and hospital
}
\description{
lage vekttapdetaljer
}
