% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWL.R
\name{GB20}
\alias{GB20}
\title{Pick particular hospitals and years from a data frame}
\usage{
GB20(sh, yr)
}
\arguments{
\item{sh}{hospital}

\item{yr}{year(s)}
}
\value{
A data frame for choice of hospitals and years
}
\description{
Pick particular hospitals and years from a data frame
}
