% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\alias{get_allevarnum}
\title{Get SoReg data}
\usage{
get_allevarnum(registry_name)
}
\arguments{
\item{registry_name}{Character string providing the registry name}
}
\value{
A data frame of registry data
}
\description{
Functions used to get data from SoReg database
}
