% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{Get_Hospital_Name}
\alias{Get_Hospital_Name}
\title{Fra RESH til Sykehusnavn}
\usage{
Get_Hospital_Name(df, RESHId)
}
\arguments{
\item{df}{data.table}

\item{RESHId}{num}
}
\value{
dataframe tabell
}
\description{
Fra RESH til Sykehusnavn
}
