% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegData.R, R/LagDD.R
\name{getRegData}
\alias{getRegData}
\alias{getRegDataWhat}
\alias{getDataDump}
\alias{lagDDump}
\title{Provide dataframe of registry data}
\usage{
getRegDataWhat(registryName)

getDataDump(registryName, tableName, fromDate, toDate, ...)

getRegDataWhat(registryName)

lagDDump(registryName, tableName, fromDate, toDate, ...)
}
\arguments{
\item{registryName}{String providing the current registryName}

\item{tableName}{String providing a table name}

\item{fromDate}{String providing start date}

\item{toDate}{String provideing end date}

\item{...}{Optional arguments to be passed to the function}
}
\description{
Provides a dataframe containing data from a registry

Provides a dataframe containing data from a registry
}
