% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aarsktrl.R
\name{k1}
\alias{k1}
\title{Pick particular hospitals and years from a data frame}
\usage{
k1(sh, yr)
}
\arguments{
\item{sh}{hospital}

\item{yr}{year(s)}
}
\value{
k1 A data frame for choice of hospitals and years
}
\description{
Pick particular hospitals and years from a data frame
}
