% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aarsktrl.R
\name{k2}
\alias{k2}
\title{Pick particular hospitals and years from a data frame}
\usage{
k2(sh, yr)
}
\arguments{
\item{sh}{hospital}

\item{yr}{year(s)}
}
\value{
k2 A data frame for choice of hospitals and years
}
\description{
Pick particular hospitals and years from a data frame
}
