% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki.R
\name{ki}
\alias{ki}
\title{Calculate indicator}
\usage{
ki(df, indicator)
}
\arguments{
\item{df}{Data frame holding SoReg data}

\item{indicator}{Character string defining the indicator to be returned.
One of "liggetid", "kompl", "dag30" or "TWL20"}
}
\value{
A tibble providing the numerator (cases), denominator (sample size)
and the indicator
}
\description{
Calculate indicator
}
