% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_funksjoner.R
\name{ki_funksjoner}
\alias{ki_funksjoner}
\alias{ki_liggetid}
\alias{ki_30dager}
\alias{ki_kompl_alv}
\title{Provide quality indicator}
\usage{
ki_liggetid(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
res tibble
}
\description{
Provides a tibble
}
