% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{kompl_gr}
\alias{kompl_gr}
\title{lage komplikasjonsgraf}
\usage{
kompl_gr(df)
}
\arguments{
\item{df}{data frame}
}
\value{
df data frame grouped by year and hospital
}
\description{
lage komplikasjonsgraf
}
