% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{kompl_tb}
\alias{kompl_tb}
\title{lage komplikasjontabell}
\usage{
kompl_tb(df, agg)
}
\arguments{
\item{df}{data frame}

\item{agg}{lgl}
}
\value{
df data frame grouped by year and hospital
}
\description{
lage komplikasjontabell
}
