% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{nitti}
\alias{nitti}
\title{slingringsmonn for aarskontrollar, interval}
\usage{
nitti(yr = 1, dag = lubridate::today(), l = 90)
}
\arguments{
\item{yr}{years}

\item{dag}{operationday}

\item{l}{number of deviating controldays}
}
\value{
interval of 'normtid'
}
\description{
slingringsmonn for aarskontrollar, interval
}
