% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{nitti_m}
\alias{nitti_m}
\title{slingringsmonn for aarskontrollar, minus}
\usage{
nitti_m(yr = 1, dag = lubridate::today(), l = 90)
}
\arguments{
\item{yr}{years}

\item{dag}{operationday}

\item{l}{number of deviating controldays}
}
\value{
beginning day of 'normtid'
}
\description{
+-90
}
