% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{nitti_p}
\alias{nitti_p}
\title{slingringsmonn for aarskontrollar, plus}
\usage{
nitti_p(yr = 1, dag = lubridate::today(), l = 90)
}
\arguments{
\item{yr}{years}

\item{dag}{operationday}

\item{l}{number of deviating controldays}
}
\value{
end day of 'normtid'
}
\description{
slingringsmonn for aarskontrollar, plus
}
