% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.data}
\alias{.data}
\alias{.env}
\title{Data pronouns}
\usage{
.data

.env
}
\description{
See \code{rlang::\link[rlang:tidyeval-data]{.data}} for details

See \code{rlang::\link[rlang:tidyeval-data]{.env}} for details
}
\keyword{internal}
