% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{reinn_tb}
\alias{reinn_tb}
\title{lage reinnleggningtabell}
\usage{
reinn_tb(df, agg)
}
\arguments{
\item{df}{data frame}

\item{agg}{lgl}
}
\value{
df data frame grouped by year and hospital
}
\description{
lage reinnleggningtabell
}
