% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{RESH_to_sh}
\alias{RESH_to_sh}
\title{Fra RESH til Sykehusnavn}
\usage{
RESH_to_sh(df, RESHId)
}
\arguments{
\item{df}{data.table}

\item{RESHId}{num}
}
\value{
dataframe tabell
}
\description{
Fra RESH til Sykehusnavn
}
