% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{siivu}
\alias{siivu}
\title{Pick particular hospitals and years from a data frame}
\usage{
siivu(df, sh, yr, prm, opr, oa)
}
\arguments{
\item{df}{Data frame holding SoReg data}

\item{sh}{hospital(s)}

\item{yr}{year(s)}

\item{prm}{primary operation}

\item{opr}{operation type}

\item{oa}{one anastomosis GB}

\item{dayIv}{dato interval}
}
\value{
A data frame for choice of hospitals, years and opr.type
}
\description{
Pick particular hospitals and years from a data frame
}
