% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{slice}
\alias{slice}
\title{Pick particular hospitals and years from a data frame}
\usage{
slice(df, sh, yr, prm, opr)
}
\arguments{
\item{df}{Data frame holding SoReg data}

\item{sh}{hospital(s)}

\item{yr}{year(s)}

\item{prm}{primary operation}

\item{opr}{operation type}
}
\value{
A data frame for choice of hospitals, years and opr.type
}
\description{
Pick particular hospitals and years from a data frame
}
