% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{snitt}
\alias{snitt}
\title{Pick particular hospitals and years from a data frame}
\usage{
snitt(df, sh, yr)
}
\arguments{
\item{df}{Data frame holding SoReg data}

\item{sh}{hospital(s)}

\item{yr}{year(s)}
}
\value{
A data frame for choice of hospitals and years
}
\description{
Pick particular hospitals and years from a data frame
}
