% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWL.R
\name{TWL}
\alias{TWL}
\title{Pick particular hospitals and years from a data frame}
\usage{
TWL(dt)
}
\arguments{
\item{dt}{Data frame holding full SoReg data}
}
\value{
A data frame for choice of hospitals and years

A data frame for choice of hospitals and years
}
\description{
Pick particular hospitals and years from a data frame
}
