% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{TWL_tb}
\alias{TWL_tb}
\title{lage vekttaptabell}
\usage{
TWL_tb(df, opr_tp)
}
\arguments{
\item{df}{data frame full data}

\item{opr_tp}{operasjonstype sleeve, bypass, oagb}
}
\value{
df data frame grouped by year and hospital
}
\description{
lage vekttaptabell
}
