% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{twlTb}
\alias{twlTb}
\title{lage vekttaptabell}
\usage{
twlTb(df, opr_tp, opr_oa = 2)
}
\arguments{
\item{df}{data frame full data}

\item{opr_tp}{operasjonstype sleeve, bypass}

\item{opr_oa}{oagb}
}
\value{
df data frame grouped by year and hospital
}
\description{
lage vekttaptabell
}
