% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{aar_ktr_gr}
\alias{aar_ktr_gr}
\title{#' lage aarskontrolltabell
#' @param df data frame
#' @param k which year control
#' @return df data frame grouped by year and hospital
#' @export}
\usage{
aar_ktr_gr(df, k)
}
\arguments{
\item{df}{data frame}

\item{k}{which year control}
}
\value{
df data frame grouped by year and hospital
}
\description{

}
\details{

}
