% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitt.R
\name{detail}
\alias{detail}
\title{lage vekttapdetaljer}
\usage{
detail(dm, agg)
}
\arguments{
\item{dm}{data frame}

\item{agg}{lgl resulater aggregert}
}
\value{
df data frame grouped by year and hospital
}
\description{
lage vekttapdetaljer
}
