% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\alias{data_years}
\alias{get_allevarnum}
\alias{get_arsrp}
\alias{describe_db}
\title{Get SoReg data}
\usage{
data_years(registry_name)

get_allevarnum(registry_name)

get_arsrp(registry_name)

describe_db(registry_name, tabs = c())
}
\arguments{
\item{registry_name}{Character string providing the registry name}

\item{tabs}{Character vector where each element represents a database
table}
}
\value{
A data frame of registry data
}
\description{
Functions used to get data from SoReg database
}
